25_python-sql/catalog/pyproject.toml
[project]
name = "exercise"
version = "0.1.0"
description = "Exercise"
authors = "Hexlet"
requires-python = ">=3.13"
dependencies = [
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
]
